Move Workflow classes in Model directory
