+++ src/Activity/Activity.php
@@ -2,7 +2,7 @@
 
 namespace Phlow\Activity;
 
-use Phlow\Model\Workflow\WorkflowNode;
+use Phlow\Model\WorkflowNode;
 
 /**
  * Interface Activity
