+++ src/Engine/WorkflowInstance.php
@@ -9,8 +9,8 @@ use Phlow\Handler\TaskHandler;
 use Phlow\Event\EndEvent;
 use Phlow\Event\StartEvent;
 use Phlow\Gateway\ExclusiveGateway;
-use Phlow\Model\Workflow\Workflow;
-use Phlow\Model\Workflow\WorkflowNode;
+use Phlow\Model\Workflow;
+use Phlow\Model\WorkflowNode;
 
 /**
  * Class WorkflowInstance
