+++ src/Model/NotFoundException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phlow\Model\Workflow;
+namespace Phlow\Model;
 
 class NotFoundException extends \Exception
 {
