+++ src/Model/WorkflowNode.php
@@ -1,5 +1,5 @@
 <?php
-namespace Phlow\Model\Workflow;
+namespace Phlow\Model;
 
 /**
  * Interface WorkflowNode
