+++ tests/Engine/SingleConnectionHandlerTest.php
@@ -5,7 +5,7 @@ namespace Phlow\Tests\Engine;
 use Phlow\Activity\Task;
 use Phlow\Engine\Exchange;
 use Phlow\Handler\SingleConnectionHandler;
-use Phlow\Model\Workflow\WorkflowConnection;
+use Phlow\Model\WorkflowConnection;
 use PHPUnit\Framework\TestCase;
 
 class SingleConnectionHandlerTest extends TestCase
